[https://issues.apache.org/jira/browse/NPANDAY-186]
Remove the uac and pab directories

Intial fix for the removal of the uac and pab directories. 

the output is already in the maven covention default repository.

however there is a problem with the npanday.plugin.runner.exe 

since its unable to use the target directory as the holder of dependencies.

git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/branches/npanday-uac-removed@1005349 13f79535-47bb-0310-9956-ffa450edef68