Exclude Silverlight SDK libraries from the POM
This was causing the unit tests to fail as soon as System.Windows is installed
in the GAC as it generates an additional element, and was causing the wrong
System.Net reference to be included in POMs previously. These libraries differ
from the ones provided by the SDK that MSBuild refers to and are used at
runtime.


git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1442450 13f79535-47bb-0310-9956-ffa450edef68