[NPANDAY-578] add support for target framework directory
Added via a property so that the build system can easily override. Using
system paths are certainly not ideal, but this may encourage users to more
easily load the artifacts into a suitable repository rather than the
effective system path of the GAC. In addition, the property paves the way to
be able to set that at build time from NPanday's Maven plugins using a
similar resolution technique.


git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1451803 13f79535-47bb-0310-9956-ffa450edef68