[NPANDAY-480] allow mapping dependency role artifact names
Some roles may carry a different artifact ID to the name that needs to be assigned to the role.
One particular case to consider is a Silverlight web application that uses RIA services - it
requires that the msdeploy section be split out into a different project to avoid creating a
circular dependency - however Azure needs the original role name of the web project not the
packaging project that it consumes.


git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1452714 13f79535-47bb-0310-9956-ffa450edef68