[NPANDAY-579] Discover referenced assemblies
When generating dependencies, we need to also inspect the referenced
assemblies.  Currently, we look for those that exist in the same directory
as the assembly and add them to the POM if they exist, as well as to the
generated POM for the assembly copied to the local Maven repository.

git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1513126 13f79535-47bb-0310-9956-ffa450edef68