[NPANDAY-630] remove duplicate asp lifecycle
This is already defined (differently!) in the dotnet-core. The one in the
plugin is incorrect, as it doesn't include compilation. This was ignored
in Maven 2, but is being picked up in Maven 3. Leave it in the core since
not all asp plugins might declare the aspx plugin with extensions on.

git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1609674 13f79535-47bb-0310-9956-ffa450edef68