[NPANDAY-505] consolidate XSD definition again
It should be ok to use a newer Windows SDK in most cases, rather than
requiring several old versions to target those builds. The compiler will
still produce binaries for the right framework. Keep these consistent for
now, until we have a richer way of defining the SDKs to use.

git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1609895 13f79535-47bb-0310-9956-ffa450edef68