[NPANDAY-598] re-instate previous fix
This is probably not the ideal way to address this issue - it seems like
the "cache" are actually collectors that should be scoped to the
resolution request, and are storing instead of saving computation. The
resolve cache is safe as it is only used to intersect with the resolved
artifacts, but the dependency cache will grow as highlighted in this
issue. These could be adjusted in future as a listener, or a processor on
the resolved artifacts, or something passed in to the resolution request,
allowing these to become singleton objects again.

git-svn-id: https://svn.apache.org/repos/asf/incubator/npanday/trunk@1619669 13f79535-47bb-0310-9956-ffa450edef68